back-end-practice/backend/pfnBackend/requirements.txt
Django
django-filter
djangorestframework
Markdown